+++ packages/i18n/src/locale/en.ts
@@ -4,6 +4,7 @@ import { common } from "./module/common";
 import { connector } from "./module/connector";
 import { leverage } from "./module/leverage";
 import { markets } from "./module/markets";
+import { navigation } from "./module/navigation";
 import { notification } from "./module/notification";
 import { orderEntry } from "./module/orderEntry";
 import { orders } from "./module/orders";
@@ -14,7 +15,6 @@ import { share } from "./module/share";
 import { tpsl } from "./module/tpsl";
 import { trading } from "./module/trading";
 import { tradingLeaderboard } from "./module/tradingLeaderboard";
-import { tradingPoints } from "./module/tradingPoints";
 import { tradingRewards } from "./module/tradingRewards";
 import { tradingView } from "./module/tradingView";
 import { transfer } from "./module/transfer";
@@ -42,8 +42,8 @@ export const en = {
   ...affiliate,
   ...ui,
   ...tradingLeaderboard,
-  ...tradingPoints,
   ...widget,
   ...vaults,
   ...notification,
+  ...navigation,
 };