+++ packages/ui-tradingview/src/tradingviewAdapter/hooks/useCreateRenderer.ts
@@ -96,6 +96,8 @@ export default function useCreateRenderer(
         unrealPnlDecimal: 2,
         basePriceDecimal: 4,
         marginMode: item.margin_mode,
+        margin_mode: item.margin_mode,
+        markPrice: item.mark_price,
       };
     });
     renderer?.renderPositions(positionList);