+++ packages/perp/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @orderly.network/futures
 
+## 4.12.1
+
+### Patch Changes
+
+- publish
+- 75b87bc: publish
+- Updated dependencies
+- Updated dependencies [75b87bc]
+  - @orderly.network/types@2.12.1
+  - @orderly.network/utils@2.12.1
+
+## 4.12.1-alpha.0
+
+### Patch Changes
+
+- publish
+- Updated dependencies
+  - @orderly.network/types@2.12.1-alpha.0
+  - @orderly.network/utils@2.12.1-alpha.0
+
 ## 4.12.0
 
 ### Minor Changes