+++ packages/portfolio/CHANGELOG.md
@@ -1,5 +1,53 @@
 # @orderly.network/portfolio
 
+## 2.12.1
+
+### Patch Changes
+
+- publish
+- 75b87bc: publish
+- Updated dependencies
+- Updated dependencies [75b87bc]
+  - @orderly.network/react-app@2.12.1
+  - @orderly.network/chart@2.12.1
+  - @orderly.network/hooks@2.12.1
+  - @orderly.network/i18n@2.12.1
+  - @orderly.network/perp@4.12.1
+  - @orderly.network/types@2.12.1
+  - @orderly.network/ui@2.12.1
+  - @orderly.network/ui-chain-selector@2.12.1
+  - @orderly.network/ui-connector@2.12.1
+  - @orderly.network/ui-leverage@2.12.1
+  - @orderly.network/ui-orders@2.12.1
+  - @orderly.network/ui-positions@2.12.1
+  - @orderly.network/ui-scaffold@2.12.1
+  - @orderly.network/ui-share@2.12.1
+  - @orderly.network/ui-transfer@2.12.1
+  - @orderly.network/utils@2.12.1
+
+## 2.12.1-alpha.0
+
+### Patch Changes
+
+- publish
+- Updated dependencies
+  - @orderly.network/react-app@2.12.1-alpha.0
+  - @orderly.network/chart@2.12.1-alpha.0
+  - @orderly.network/hooks@2.12.1-alpha.0
+  - @orderly.network/i18n@2.12.1-alpha.0
+  - @orderly.network/perp@4.12.1-alpha.0
+  - @orderly.network/types@2.12.1-alpha.0
+  - @orderly.network/ui@2.12.1-alpha.0
+  - @orderly.network/ui-chain-selector@2.12.1-alpha.0
+  - @orderly.network/ui-connector@2.12.1-alpha.0
+  - @orderly.network/ui-leverage@2.12.1-alpha.0
+  - @orderly.network/ui-orders@2.12.1-alpha.0
+  - @orderly.network/ui-positions@2.12.1-alpha.0
+  - @orderly.network/ui-scaffold@2.12.1-alpha.0
+  - @orderly.network/ui-share@2.12.1-alpha.0
+  - @orderly.network/ui-transfer@2.12.1-alpha.0
+  - @orderly.network/utils@2.12.1-alpha.0
+
 ## 2.12.0
 
 ### Minor Changes