+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @orderly.network/ui
 
+## 2.12.1
+
+### Patch Changes
+
+- publish
+- 75b87bc: publish
+- Updated dependencies
+- Updated dependencies [75b87bc]
+  - @orderly.network/utils@2.12.1
+
+## 2.12.1-alpha.0
+
+### Patch Changes
+
+- publish
+- Updated dependencies
+  - @orderly.network/utils@2.12.1-alpha.0
+
 ## 2.12.0
 
 ### Minor Changes